Rename the core project
