+++ pom.xml
@@ -7,7 +7,7 @@
 	<version>0.0.1-SNAPSHOT</version>
     <packaging>pom</packaging>
     <modules>
-        <module>streamloots</module>
+        <module>streamlootscore</module>
         <module>streamlootsbase</module>
     </modules>
 </project>
