+++ streamlootsbase/dependency-reduced-pom.xml
@@ -55,7 +55,7 @@
     </dependency>
     <dependency>
       <groupId>fr.entrivax</groupId>
-      <artifactId>streamloots</artifactId>
+      <artifactId>streamlootscore</artifactId>
       <version>0.0.1-SNAPSHOT</version>
       <scope>provided</scope>
     </dependency>
