+++ streamlootsbase/pom.xml
@@ -28,7 +28,7 @@
 		</dependency>
 		<dependency>
 			<groupId>fr.entrivax</groupId>
-			<artifactId>streamloots</artifactId>
+			<artifactId>streamlootscore</artifactId>
 			<version>0.0.1-SNAPSHOT</version>
 			<scope>provided</scope>
 		</dependency>
