+++ streamlootsbase/src/main/java/fr/entrivax/streamlootsbase/commands/StreamlootsCardGiveItemCommand.java
@@ -11,7 +11,7 @@ import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.plugin.java.JavaPlugin;
 
-import fr.entrivax.streamloots.commands.IStreamlootsCardCommand;
+import fr.entrivax.streamlootscore.commands.IStreamlootsCardCommand;
 import fr.entrivax.streamlootsbase.PlayersHelper;
 
 public class StreamlootsCardGiveItemCommand implements IStreamlootsCardCommand {
