+++ streamlootsbase/src/main/java/fr/entrivax/streamlootsbase/commands/StreamlootsCardSetHealthCommand.java
@@ -9,7 +9,7 @@ import org.bukkit.attribute.Attribute;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.java.JavaPlugin;
 
-import fr.entrivax.streamloots.commands.IStreamlootsCardCommand;
+import fr.entrivax.streamlootscore.commands.IStreamlootsCardCommand;
 import fr.entrivax.streamlootsbase.PlayersHelper;
 
 public class StreamlootsCardSetHealthCommand implements IStreamlootsCardCommand {
