+++ streamlootscore/.project
