+++ streamlootscore/.settings/org.eclipse.jdt.apt.core.prefs
