+++ streamlootscore/.settings/org.eclipse.jdt.core.prefs
