+++ streamlootscore/.settings/org.eclipse.m2e.core.prefs
