+++ streamlootscore/LICENSE
