+++ streamlootscore/src/main/java/com/cisco/commons/networking/EventHandler.java
