+++ streamlootscore/src/main/java/com/cisco/commons/networking/SSEClient.java
