+++ streamlootscore/src/main/java/fr/entrivax/streamlootscore/CardConfig.java
@@ -1,4 +1,4 @@
-package fr.entrivax.streamloots;
+package fr.entrivax.streamlootscore;
 
 import java.util.List;
