+++ streamlootscore/src/main/java/fr/entrivax/streamlootscore/commands/StreamlootsCardRegistry.java
@@ -1,4 +1,4 @@
-package fr.entrivax.streamloots.commands;
+package fr.entrivax.streamlootscore.commands;
 
 import java.util.HashMap;
