+++ streamlootscore/src/main/resources/config.yml
